src/main/webapp/app/entities/config/update/config-update.component.spec.ts
import {ComponentFixture, TestBed} from '@angular/core/testing';
import {HttpResponse} from '@angular/common/http';
import {HttpClientTestingModule} from '@angular/common/http/testing';
import {FormBuilder} from '@angular/forms';
import {ActivatedRoute} from '@angular/router';
import {RouterTestingModule} from '@angular/router/testing';
import {from, of, Subject} from 'rxjs';

import {ConfigService} from '../service/config.service';
import {Config, IConfig} from '../config.model';

import {ConfigUpdateComponent} from './config-update.component';

describe('Config Management Update Component', () => {
  let comp: ConfigUpdateComponent;
  let fixture: ComponentFixture<ConfigUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let configService: ConfigService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [ConfigUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(ConfigUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(ConfigUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    configService = TestBed.inject(ConfigService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should update editForm', () => {
      const config: IConfig = { id: 456 };

      activatedRoute.data = of({ config });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(config));
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Config>>();
      const config = { id: 123 };
      jest.spyOn(configService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ config });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: config }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(configService.update).toHaveBeenCalledWith(config);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Config>>();
      const config = new Config();
      jest.spyOn(configService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ config });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: config }));
      saveSubject.complete();

      // THEN
      expect(configService.create).toHaveBeenCalledWith(config);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Config>>();
      const config = { id: 123 };
      jest.spyOn(configService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ config });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(configService.update).toHaveBeenCalledWith(config);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });
});
